Modified handling of stacked contacts to fail a route rather than accept it if a solution cannot be found that avoids stacking contacts beyond what is specified as allowable.  In stage 2, a stacked contact solution will always be found, by overrouting existing nets and moving those nets to the rip-up list.
Also modified behavior to specify a maximum amount of attempts
to rip-up and re-route nets, to avoid an infinite loop when
routing an over-congested area.  At the end, failed and abandoned
routes are printed to the terminal as well as the "failed" file,
so that they are immediately obvious in the output.